Refactored app.rb and main.rb files
