+++ add_book.rb
@@ -1,0 +1,8 @@
+def add_book(books)
+    print 'Title: '
+    title = gets.chomp
+    print 'Author: '
+    author = gets.chomp
+    books << Book.new(title, author)
+    puts 'Book created successfully.'
+end
